qr-project-main/requirements.txt
Flask
opencv-python-headless
qrcode[pil]
Pillow
python-multipart
cairosvg
